backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.0
pydantic==2.5.2
pydantic-settings==2.1.0
redis==5.0.1
celery==5.3.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 邮件服务依赖
aiosmtplib==3.0.1
email-validator==2.1.0
jinja2==3.1.2

# 测试依赖
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0
aiosqlite==0.19.0
